// frontend-app/src/apiClient/models/TokenHolderBalanceUpdateType.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export enum TokenHolderBalanceUpdateType {
	MINT = "Mint",
	BURN = "Burn",
	TRANSFER_OUT = "TransferOut",
	TRANSFER_IN = "TransferIn",
	FREEZE = "Freeze",
	UN_FREEZE = "UnFreeze",
}