// frontend-app/src/apiClient/models/PagedResponse_TokenHolderUserBalanceUpdate.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { TokenHolderUserBalanceUpdate } from "./TokenHolderUserBalanceUpdate";

/**
 * Pages Response. This is a generic response that can be used to return a list
 * of items with pagination.
 */
export type PagedResponse_TokenHolderUserBalanceUpdate = {
	page_count: number;
	page: number;
	data: Array<TokenHolderUserBalanceUpdate>;
};